Rename args to more specific terms
While it's correct to name them "source" and "target" to indicate
the direction of the actions about to take plae, they are generic
terms. And "source directory" might even be mistaken for "source code
directory". The more specific term is "install prefix" as the
source (of the unbundled application), especially since that's what
the description is already calling it.

"output directory" on the other hand is not an improvment over
"target directory" but allows the args to be "i" and "o", leaning
on the common pattern of "input" and "output".